+++ packages/css-color-parser/src/functions/lch-normalize-channel-values.ts
@@ -41,7 +41,7 @@ export function normalize_LCH_ChannelValues(token: CSSToken, index: number, colo
 
 		let value = normalize(token[4].value, 1, 0, 100);
 		if (index === 1) {
-			value = normalize(token[4].value, (100 / 150), 0, Infinity);
+			value = normalize(token[4].value, (100 / 150), 0, 2_147_483_647);
 		} else if (index === 3) {
 			value = normalize(token[4].value, 100, 0, 1);
 		}
@@ -65,7 +65,7 @@ export function normalize_LCH_ChannelValues(token: CSSToken, index: number, colo
 
 		let value = normalize(token[4].value, 1, 0, 100);
 		if (index === 1) {
-			value = normalize(token[4].value, 1, 0, Infinity);
+			value = normalize(token[4].value, 1, 0, 2_147_483_647);
 		} else if (index === 3) {
 			value = normalize(token[4].value, 1, 0, 1);
 		}