+++ packages/css-color-parser/src/functions/oklab-normalize-channel-values.ts
@@ -27,7 +27,7 @@ export function normalize_OKLab_ChannelValues(token: CSSToken, index: number, co
 
 		let value = normalize(token[4].value, 100, 0, 1);
 		if (index === 1 || index === 2) {
-			value = normalize(token[4].value, 250, -Infinity, Infinity);
+			value = normalize(token[4].value, 250, -2_147_483_647, 2_147_483_647);
 		} else if (index === 3) {
 			value = normalize(token[4].value, 100, 0, 1);
 		}
@@ -51,7 +51,7 @@ export function normalize_OKLab_ChannelValues(token: CSSToken, index: number, co
 
 		let value = normalize(token[4].value, 1, 0, 1);
 		if (index === 1 || index === 2) {
-			value = normalize(token[4].value, 1, -Infinity, Infinity);
+			value = normalize(token[4].value, 1, -2_147_483_647, 2_147_483_647);
 		} else if (index === 3) {
 			value = normalize(token[4].value, 1, 0, 1);
 		}