+++ packages/css-calc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow uncapped precision in `calc()` serialization
 - Add `censorIntoStandardRepresentableValues` conversion option 
+- Updated [`@csstools/css-tokenizer`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer) to [`2.3.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer/CHANGELOG.md#230) (minor)
 
 ### 1.2.0